Add function for summing numbers in range
Use if/else condition to run the function only if a and b are non-negative integers, otherwise it returns 'ERROR'. a and b are first added to a variable titled array, then sorted so that the function will work even if a is larger. The function runs a for loop to add all numbers in the range to a fullArray variable, and then sums them using the forEach method.